+++ src/contents/config/config.qml
@@ -29,6 +29,11 @@ ConfigModel {
         source: "configNotifications.qml"
     }
     ConfigCategory {
+        name: i18n("Printer states")
+        icon: "printer"
+        source: "configPrinter.qml"
+    }
+    ConfigCategory {
         name: i18n("OctoPrint API")
         icon: "configure"
         source: "configApi.qml"