+++ CHANGES.md
@@ -12,6 +12,7 @@
    * Compact widget's tooltip now shows print and remaining time.
    * Compact widget can now display elapsed and remaining print time.
    * Visibility of compact widget state icon can be configured per state bucket.
+   * Printer status bucket names can be customiezed.
 
 * v1.1.1 (2020-12-03)
    * Improved docs